Failing test case for hyperlinked generic related fields.
If the correct view_name is given to the ManyHyperlinkedRelatedField
('bookmarks-detail'), the test passes, though of course this only
works because the test relationship is a Bookmark. When it does work,
the returned URLs are not absolute ("http://testserver/..."). I
assume this is because no real request/response context exists.