pass many=False to pagination_serializer_class()
pagination serializers should always have many=False set, otherwise they'll
be passed the individual objects of the page rather than the page object
itself. this is a real/better fix for #1186. removes the need for the
deprecated isinstance Page check, though it's left in place since it
doesn't hurt anything.