make serializers more extensible
The general idea of this commit is to be able to pass metadata about
the serialization to the serializers, so the serializers can change
their behavior.

In order to do this, we must be able to (1) pass additional
parameters to the serializer at serialization time, (2) the
serializer must be able to do something with those parameters, and
(3) the serializer must be able to pass the parameters down to
child serializers.

(1) to_native now accepts *args, **kwargs
(2) we add several extension points to to_native so subclasses can
do things with the *args, **kwargs. to_native now calls _skip_field
to determine whether a given field should be skipped. A subclass
can subclass this method to skip certain fields. to_native
now calls _get_field_value to actually get a subfield's value.
_get_field_value can be subclassed to modify the *args, *kwargs
passed to a child serializer, or to modify the child field's value
based on the *args,**kwargs.
(3) to_native calls child fields with the *args, **kwargs. However
it does so in a try/except block so if it is an old field without
support for *args, **kwargs, it will not fail, it simply won't
support the *args,**kwargs functionality. field_to_native now
passes all *args, **kwargs received into to_native.

all tests pass