rework Field.__init__ to use either kwargs / class attrs.
Precedence is given to kwargs over class attributes.
You can create a subclass of a Field for 99% of your use cases while still
needing a one-off instance where you need to change an attribute; hence
kwargs is still useful.

Other important point is that kwargs is the current de-facto standard for
declaring fields in a serializer.