Pass a dict as data to django-filter
In order to filter with multivalue params such as 'in' lookups there are two alternatives:

* ?multiparam=1,2,3
* ?multiparam=1&multiparam=2

The second option is a better approach when facing strings. However, by definition QueryDict does only take into account the last param included.

Django-filters expect a dict as data. To be congruent with what is passed in the request we need to transform `query_params` to a `dict`.

As can be seen in the test written, adhoc programming must be done also in django-filter to allow 'in' lookups but that's another issue.

Part of the solution was taken from @leo-the-manic in http://stackoverflow.com/questions/13349573/how-to-change-a-django-querydict-to-python-dict#answer-22100334