provide get_success_headers on UpdateModelMixin
This change provides a `get_success_headers` method on the UpdateModelMixin which can be used like the corresponding method on the CreateModelMixin. I thought this might be usefull because I was trying to add a header based on the result of the update operation, and this seems like a valid way.