Added max_length, min_length keyword arguments to ListField
Sometimes it's useful to reject lists if they are too long.
E.g. in performance purposes, if you make a database request afterwords,
you would prefer not having a request that pulls out thousands of
ORM objects after all the validations had passed.

And the minimum length I added just to have it there. Don't see
any reason to have it for now.