Added view name to has_permission error message
When this error occurs in a project with a large number of views, it is difficult to pinpoint in which view this error might be happening. By explicitly including the view name in the error message the developer can more easily identify the cause of the problem. Without including this the error message is close to useless as it requires source code analysis or Googling. There are many examples of these kinds of errors in DRF and Django itself, but we have to start somewhere.