Handle empty body with json content-type as json
An empty request body should not imply HTML form semantics.

An attribute in a JSON object can either be:
 - not set at all,
 - null,
 - an empty list.

An HTML form isn't as flexible, so invariably turning an empty body into
a QueryDict makes it impossible to determine if a given value was passed
in as an empty list or if it was not set at all.