Fix multiple value bug in html.parse_html_dict()
It is possible that a key in a MultiValueDict has multiple values, lists
are represented this way. When accessing a key in a MultiValueDict
it only returns the last element of that key. This becomes a problem
when parsing an html dict with a list inside of it.

To fix this problem we have to get and set the value using .getlist()
and .setlist().

This commit solves a bug that caused a list in a nested serializer to
not being properly parsed if it was sent as multipart formated data.