adds test for raising uncaught exc in serializer
Serializers should certainly except `ValidationError`s and respond
with the appropriate `400 Bad Request`.

This commit adds a test to check that it remains possible to raise
another exception - e.g. deriving `APIException` - that escapes
the serializer and is handled elsewhere - or not, as the case may
be.

It turns out that this behaviour is already supported, but it
wasn't clear that it was (or was by-design) from reading the
existing tests; while it _seemed_ to be the case from the source,
it's not as obvious as it would be if there's an explicit test.