Changed RequestsClient to accept an underlying client
The RequestsClient uses an underlying client, by default APIClient,
to perform the request. This is important because using the previous
version of RequestsClient did not integrate well into tests which
previously relied on the django.test.client.Client behavior.

For example, making a query after making a request would fail due
to the response signal handlers closing the database connection.