modify _writable_fields in serializers.py
Sometimes 'self' instance is not a field (used in rest_framework)
If the 'serializer' that contains 'field' is 'self', only the subelements of
the serializer, the 'fields', are returned.
in the case of a serializer that contains 'field' as a 'child' element,
Only instances of `.source` ending with` _set '(following relationships
"backward") are found using regular expressions.
Then use the child elements again to find the field using '_writable_fields'.

Do not modify '_readable_fields' because it does not address the issue of what
is currently being viewed.