Modify 'to_internal_value' method for handling list
modified the 'get_value' method in fields.py,
The value of primitive_value is a list of values to match the field.

Validate the primitive_value (list) elements one by one using the
run_validation method.

The verified value is appended to the validated_list

Finally, only when there are two or more elements of validated_list, they are
returned as a list.