Custom trailing_slash
If a string is passed into the SimpleRouter, append that to the regex determining the route. This will prevent unnecessary confusion when a PATCH/PUT request is made to a given endpoint and a slash is omitted. Have lost a bit of dev time figuring out where the `request.data` has gone because the endpoint was right, but the slash was missing.

With this change, one can use  `SimpleRouter(trailing_slash='/?')` to direct to any of the given endpoints, while not losing any of the prior functionality.