Fix throttling documentation about Remote-Addr
There's no such HTTP header called Remote-Addr and the throttling code is indeed not using any HTTP variable to get this.
Instead it is using the WSGI environment variable REMOTE_ADDR, which may be set by the webserver according to configuration. For example, in nginx it is possible to use the nginx_http_realip_module and set it based on any HTTP header, if it comes from a trusted reverse proxy server.

On the other hand, X-Forwarded-For is a valid HTTP header and that's why it is also prefixed with HTTP_ in the request.META object.