Example code don't check for non existing fields
The behavior of DynamicFieldsModelSerializer when passing a non-existent (with a typo for example) field is confusing:
  * The field will not appear in the serializer output as expected
  * The error isn't caught anywhere (which can be frustrating to debug)

This assert takes care of this. Maybe there is a better way to throw an exception than this, feedback welcome. 

BTW, I could imagine having this functionality in a mixin. What do you think about it? Let me know