Fix support for ListField default values in HTML mode
The ListField handles html forms data differently than json data.  This section
would always return a value (an empty list) for a list field, even when that
field name was not posted at all.  This change checks for the field being
present and returns `empty` if it is not; this lets the standard default-value and
required field handling take over.