Support serializing date/datetime values before 1900 on python 2
Python 2 does not support strftime() for datetimes prior to 1900. This
is fixed in Python 3.x but the python developers have signaled that they
do not intend to backport the fix. The module django.utils.datetime_safe
was written to workaround this issue. I've tried to use a light hand,
only using the datetime_safe functions where serialization would
otherwise throw a ValueError.