Update serializers.md
This issue seems to come up a lot on stackoverflow, and I had to read quite a bit of code to understand this. I think it's important here to point people to get_serializer_context() in views. I've tried to provide a minimum but complete example to explain get_serializer_context()