Add tests for unique validation for bulk creation in model serializers
The unique validation in model serializers only check for existing
entries on the database. If the serializer is instanced with many=True
and receives more than one item in the data argument having the same
value for a unique field, the .is_valid() method returns True, but
calling .save() will raise an IntegrityError.