Improve documentation regarding serializer class on extra actions.
The documentation does currently not describe how to use a custom
`serializer_class` with extra actions. The example uses a custom
serializer without using the seralizer_class attribute, resulting in the
generated documentation page presenting the wrong form.

The author of the feature described the usage of the `serializer_class`
kwarg in his pr #5605:
https://github.com/encode/django-rest-framework/pull/5605#issuecomment-345420265