Replace usage of BindingDict and remove it
BindingDict ensures that .bind() is called on each field. However this
is easily ensured without it.

One thing that BindingDict might have helped with is to ensure .bind()
is called even if a field is added after the fact to the dict,
externally. However such use is already broken since _readable_fields
and _writable fields use the fields and are cached. Doing so sounds like
a bad idea, anyway.