Modified `serializers.ChoiceField` behavior convert `display_name` --> `key` during deserialization and `key` --> `display_name` during serialization.
Perhaps I misunderstand the intended behavior of `fields.ChoiceField`, but I
noticed recently that the although every `ChoiceField` instance stores the full
list of choices with which it is initialized (a list of pairs of `key` and
`display_name`), I couldn't figure out when it ever uses a choice's
`display_name`. As a result, I found myself manually converting user input
values, which matched choice `display_name`s, to the underlying data values that
match the choice `key`s.

This commit modifies the behavior of `ChoiceField` instances so that they
automatically deserialize user input into the corresponding underlying data
value and do the reverse for serialization.