Allow serializer data to include related objects by object
When instantiating a serializer from cleaned data, the relational fields
have been already converted into objects, which makes the fields
invalid. Having a valid object as the value of a field should be
considered valid, even if not in the prescribed format.

As such, this change adds a check to allow objects of the model type in
lieu of the proper representation.