Update fields.md
If I specify `blank=True` in my Model's field, then this field is not required in `ModelSerializer`.
BTW it's confusing since (as far as I know) `blank=True` has no effect on e.g. `ManyToManyField`, but when I'm using DRF then I have a reason to use `blank=True` on this field.