Add machine-readable status to Authentication classes
BasicAuthentication and TokenAuthentication have two failure cases when
they are passed technically valid credentials by an API client:

 - the passed credentials are not correct
 - the credentials are correct but the user is inactive

In both cases, only a human-readable 'detail' is returned in the 401
body, which is translated according to the site settings:

```json
HTTP/1.1 401 Unauthorized
<headers snipped>

{
    "detail": "Ungültiges Token"
}
```

The free-form text and its translation make it impossible for an API
consumer to determine the actual reason (inactive user, out of luck; or
wrong credentials, try again).

This PR adds a machine-readable 'status' field to the response, which
can take one of two values:

1. `invalid-credentials` - returned when the passed username, password
   or token were incorrect.

2. `inactive-user` - returned when the credentials were valid but the
   user account is disabled.

Example:

```json
HTTP/1.1 401 Unauthorized
<headers snipped>

{
    "detail": "Ungültiges Token",
    "status": "invalid-token"
}
```

As this only adds a machine-readable field for the already exposed
human-readable 'detail' field, there are no new security implications.